+++ src/org/meta_environment/rascal/library/experiments/VL/Use.java
@@ -22,7 +22,7 @@ public class Use extends VELEM {
 	}
 
 	@Override 
-	void bbox(int left, int top){
+	void bbox(float left, float top){
 		this.left = left;
 		this.top = top;
 		inside.bbox(left, top);
