+++ src/org/meta_environment/rascal/library/experiments/VL/VCat.java
@@ -15,7 +15,7 @@ public class VCat extends Compose {
 	}
 	
 	@Override
-	void bbox(int left, int top){
+	void bbox(float left, float top){
 		this.left = left;
 		this.top = top;
 		width = 0;
